chore: restructure code - dashboard views - learners
